+++ src/pages/reservation/api/index.ts
@@ -4,8 +4,25 @@ import type {
 } from 'apis/data-contracts';
 
 import { apiRequest, type ParamValue } from '@api/apiRequest';
-import type { FoodTrucksFilterType } from '@pages/reservation/types/food-trucks-filter-type';
+
 import { formatParams } from '@pages/reservation/utils/format-params';
+
+import type {
+  AVAILABLE_QUANTITY,
+  NEED_ELECTRICITY,
+  PAYMENT_METHOD,
+} from '@pages/filter/constant/filter-option-constants';
+
+export interface FoodTrucksFilterType {
+  regionCodes?: string[] | null;
+  schedules?: string[] | null;
+  availableQuantity?: (typeof AVAILABLE_QUANTITY)[number] | null;
+  categories?: string[] | null;
+  needElectricity?: (typeof NEED_ELECTRICITY)[number] | null;
+  paymentMethod?: (typeof PAYMENT_METHOD)[number] | null;
+
+  [key: string]: ParamValue;
+}
 
 export const getFoodTrucksData = async ({
   filter,
